requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

Click == 7.0

Flask == 1.0.2
aiohttp == 3.5.4

aioredis == 1.2.0

eth-account == 0.4.0
eth-utils == 1.6.1

flask_restplus == 0.13.0

py2neo == 4.2.0

redis == 3.2.1

requests == 2.22.0

setuptools == 41.0.1

tornado == 6.0.2

neo4j-driver == 1.6.3
neomodel == 3.3.1

flask-login == 0.4.1
sqlalchemy == 1.3.3
alembic == 1.0.10
tornado-sqlalchemy == 0.6.1
flask-cors == 3.0.7

pymysql == 0.9.3
py-bcrypt == 0.4

boto3==1.7.4

-e git+https://github.com/SPSCommerce/redlock-py.git#egg=redlock-py
aioredlock==0.3.0
tenacity==5.0.4

google-cloud-storage == 1.16.1
google-cloud-vision == 0.37.0

dynaconf == 2.0.3
flask-restx
git+https://github.com/blockvigil/ethvigil-python-sdk.git
coloredlogs